change functions parameters from int to size_t

+++ src/contract.h
@@ -17,13 +17,13 @@ typedef struct contract {
     int endOffice;
 } contract;
 
-void showContractData(contract c[], int qtd);
+void showContractData(contract c[], size_t qtd);
 void showContract(contract c);
-void editContract(contract cont[], vehicle vec[], int pos, date d[], int qtdvec, int qtd);
-void deleteContract(contract c[], int pos, int *qtd);
-int ShowContract(contract c[], date data[], int qtd);
-int searchDate(contract cont[], date data[], int qtd);
-void startContract(contract contracts[], customer customers[], vehicle vehicles[], int local[][MAX_OFFICES], date data[], int qtdcli, int qtdvec, int *qtd);
+void editContract(contract cont[], vehicle vec[], int pos, date d[], size_t qtdvec, size_t qtd);
+void deleteContract(contract c[], int pos, size_t *qtd);
+int ShowContract(contract c[], date data[], size_t qtd);
+int searchDate(contract cont[], date data[], size_t qtd);
+void startContract(contract contracts[], customer customers[], vehicle vehicles[], int local[][MAX_OFFICES], date data[], size_t qtdcli, size_t qtdvec, size_t *qtd);
 void endContract(contract contracts[], int pos, date data[], vehicle vehicles[], customer customers[], int quantityCustomers, int quantityVehicles, int quantityContracts, int local[][MAX_OFFICES]);
 
 #endif//RENT_A_CAR_CONTRACT_H
